fix : 자료형문제 test중
Former-commit-id: 2bc0ef9423899f2cdcb55975eeea421471f90b25

+++ backend/src/main/java/com/tfeo/backend/domain/home/model/entity/Home.java
@@ -46,7 +46,7 @@ public class Home {
 	private String hostName;
 
 	@Null
-	private Integer hostAge;
+	private float hostAge;
 
 	private String hostPhone;
